Grog: advancing towards item combination.
Not working yet, but patterns are compiled now for each routine.

The client issues an interaction-request-with-tool when trying to combine
an inventory item with an scene item. Not implemented by the server yet.

When the client retrieves the list of 'actions' for an item, actions with
patterns are ignored (because they are combining-actions).